[shared_preferences] Support v2 android embedding. (#2162)
