+++ packages/shared_preferences/example/android/gradle.properties
@@ -1,1 +1,4 @@
 org.gradle.jvmargs=-Xmx1536M
+android.useAndroidX=true
+android.enableJetifier=true
+android.enableR8=true
